refactor: Move provider class to provider package
