+++ viewith-server/src/main/java/com/puremadeleine/viewith/service/ReviewService.java
@@ -7,6 +7,9 @@ import com.puremadeleine.viewith.dto.review.CreateReviewReqDto;
 import com.puremadeleine.viewith.dto.review.CreateReviewResDto;
 import com.puremadeleine.viewith.dto.review.ReviewInfoResDto;
 import com.puremadeleine.viewith.dto.review.UpdateReviewReqDto;
+import com.puremadeleine.viewith.provider.ReviewProvider;
+import com.puremadeleine.viewith.provider.SeatProvider;
+import com.puremadeleine.viewith.provider.VenueProvider;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
